Avoid using dc role for gw network configuration
The current naming convention for the address gateway entry for subcloud
is 'system-controller-gateway-ip-[mgmt/admin]'. However, this format is
unnecessary as it only refers to the [mgmt/admin] network gateway. This
commit changes the name to the standard format
'controller-gateway-[mgmt/admin]', aligning it with the naming
convention used for OAM. Additionally, an upgrade script is provided to
facilitate the transition from previous versions.

[PASS] Perform a fresh install on an SX system. Configure the
management gateway and verify that the address entry is configured with
the name controller-gateway-mgmt in the addresses table of the database.

[PASS] Perform a fresh install on a DC system. Verify the names of the
management and OAM gateways in the subcloud. The names should be
configured as controller-gateway-mgmt and controller-gateway-oam,
respectively. The management gateway entry should also be present in
the addn_hosts of dnsmasq.

[PASS] Perform a fresh install on a SX IPv4/IPv6 system. Update the
management gateways in the addrpool table and verify that entries are
created in the addresses table of the database with the name
controller-gateway-mgmt. After locking and unlocking the controllers,
the IPv4 entry should be created in the addn_hosts of dnsmasq.

[PASS] Update the management gateway in the addrpool table on an SX
system. The entry should be updated in the addresses database table
with the name controller-gateway-mgmt. After locking and unlocking,
the entry should be updated in the addn_hosts of dnsmasq.

[PASS] Update the management gateway in the addrpool table on a
subcloud (dual-stack IPv4/IPv6). The entries should be updated in the
addresses database table with the name controller-gateway-mgmt. After
locking and unlocking, the IPv4 entry should be updated in the
addn_hosts of dnsmasq.

[PASS] Update the management gateway in the addrpool table on a
dual-stack IPv6/IPv4 configuration. The entries should be updated in the
addresses database table with the name controller-gateway-mgmt. After
locking and unlocking, the IPv6 entry should be updated in the
addn_hosts of dnsmasq.

[PASS] Update the OAM gateway in the addrpool table on an SX system.
The address should be correctly updated in the database and the name
should remain as controller-gateway-oam.

[PASS] Perform a fresh install of SX, DX, and subcloud systems to verify
the loopback configuration. The loopback should be correctly configured.

[PASS] Update the OAM gateway in the addrpool table on a system
controller. The address should be correctly updated in the database and
the name should remain as controller-gateway-oam.

[PASS] Create an addrpool for admin and verify that the admin gateway
address is created with the name controller-gateway-admin in the
addresses table of the database.

[PASS] Upgrade the subcloud from the previous version to the current
one. The entries in the address table of the database that have the
names 'system-controller-gateway-ip-mgmt' and
'system-controller-gateway-ip-admin' should be updated to
'controller-gateway-mgmt' and 'controller-gateway-admin', respectively.

Story: 2011312
Task: 51797

Change-Id: I9dacb44725dde5ddeb1cbcd514ef38eb433f6d2e
Signed-off-by: Ferdinando Terada <Ferdinando.GodoyTerada@windriver.com>