Implement inter-app dependency behavior support
This change adds to the app_framework the ability to handle
"action" keys with values of "ignore", "warn", and "error" from
the new dependent_apps section of the app metadata within the
manual processes of uploading, applying, removing, and deleting
an app. In a future effort, the "action" key of "apply" will also
be implemented.

Here is an example of the new structure supported in the app
metadata:
dependent_apps:
    - name: <app_name.1>
      action: <ignore|warn|error|apply>
    - name: <app_name.2>
      action: <ignore|warn|error|apply>

Here is a brief description of what each action value does:

ignore - do nothing new, react as if the dependency did not exist

warn - log a message and provide a UI status message when this
application is uploaded AND when applied that the dependent app is
not available OR not uploaded OR not applied. Application apply is
allowed to proceed as application plugins may set dynamic overrides
to accommodate. This should be the default if the 'action' is not
specified.

error - log a message and provide a UI status message when this
application is uploaded AND when applied that the dependent app is
not available OR not uploaded OR not applied. Application apply is
blocked.

apply - update the internal auto-update state of the dependent
application (if available) and apply the dependent application.
This should be paying attention to the 'behavior:
evaluate_reapply:after:[a,b,c]' ordering. Log a message and
provide a UI status message the dependent app is not available OR
not uploaded OR not applied.

Test Plan:
PASS: build-pkgs && build-image
PASS: install AIO-SX
PASS: upload, apply, remove and delete metrics-serve
PASS: upload, apply, remove and delete NFD
PASS: upload, apply, remove and delete vault

PASS: Updates the vault application with the dependent_apps key
      with one or more apps with action equal to ignore. Nothing
      changes and the upload and apply process completes without
      change as expected.

PASS: Updates the vault application with the dependent_apps key
      with one or more apps with action equal to warn. The upload
      and apply are completed successfully, but both receive a log
      message as warning and informing that there are dependent
      apps that are missing. The app's UI also receives this
      warning.

PASS: Updates the vault application with the dependent_apps key
      with one or more apps with an action equal to error. The
      upload is completed successfully, but the app's UI and logs
      report that there are missing dependent apps. The apply
      process fails because the dependencies were not previously
      installed. An error message is added to the logs. The app
      remains in apply-failed and the reason for the failure is
      added to the UI.

Story: 2011351
Task: 51843

Change-Id: Ic5216159c1fa356fcb37f6ca6384d36b75d44b05
Signed-off-by: David Bastos <david.barbosabastos@windriver.com>