Fix host runtime config check during during out-of-date audit
The cause of the error is an unhandled exception in the
out-of-date alarm audit.

This audit attempts to reapply runtime configs that cause
long lived 250.001 alarms.  However, if the 250.001 alarm is caused
by adeferred config, this will step will fail with a DB NotFound error.
This is because the config was never applied and not present in the DB yet.

To fix this, the NotFound error is now caught during the out-of-date
audit.  This allows the rest of the audit to proceed and correctly
apply the deferred config.

Another issue was noted with a TODO comment.  If sysinv-conductor is
restarted or a SWACT is triggered while deferred configs are present,
those deferred configs will be lost and the system will be stuck
out-of-date.

TEST PLAN
PASS: AIO-DX, get some deferred runtimes, ensure they work in audit
* Simplest way to get this is to slightly modify code to always defer
* Ways to trigger runtime configs:
  * system dns-modify 'nameservers=...'
  * system service-parameter-apply kubernetes

Closes-Bug: https://bugs.launchpad.net/starlingx/+bug/2102661
Change-Id: Ib1eaf9dbc3835bcbfc655757f73b3d8a29eeafab
Signed-off-by: Joshua Kraitberg <joshua.kraitberg@windriver.com>