Change Log Message to Avoid Misunderstanding
When a lab is equipped with an FEC ACCELERATOR device, such as
"intel_acc100" or "intel_acc200," and the label "sriovdp=enabled"
is present, the system reads the "pci_device" table from the sysinv
database.
An error message is generated when the field sriov_vf_pdevice_id is empty.
However, this field may be empty in many installations, and as a result,
the error message: "Failed to get device id for pci device"
often appears in the logs.
This can mislead the analysis and result in incorrect conclusions during
bug investigation.
To prevent confusion, we are changing the log level of this message from
Error to Debug.

Test done:
     AIO-SX with ACC100 FEC device install analysis

Closes-Bug: 2102642

Change-Id: I2290689b567dd3862a5a645001100eab2c6fd332
Signed-off-by: Fabiano Correa Mercer <fabiano.correamercer@windriver.com>