Conditional fix and update APPS_NOT_TO_UPDATE list
Removal of the deployment-manager application from the
APPS_NOT_TO_UPDATE list of applications, because the update of
this application can now be controlled by script 21.

A correction in the update_in_series function was also made,
preventing the script from continuing if, after the timeout,
the app is still being updated.

Test Plan:
PASS: build-pkgs && build-image
PASS: If the app does not finish updating in script 21  before 30
      checks are performed, the script should fail.
PASS: upgrade stx-10 to branch master

Closes-bug: 2080672

Change-Id: I3183adb5ba0c608012878e015456e66d929fabcd
Signed-off-by: David Bastos <david.barbosabastos@windriver.com>