Reduce manifest for reconfiguring service endpoints
The existing openstack endpoint runtime manifest not only updates the
keystone service endpoint, but also updates the service user and role
regardless whether the users and roles are changed or not. As the
OpenStack Keystone operations are slow, the
"openstack::keystone::endpoint::runtime" class is very slow to
execute, normally it takes around 5 minutes to be applied.

This commit removes the "openstack::keystone::endpoint::runtime" class
in the following cases:
1. Updating the OAM IP on AIO-SX.
2. Switch between HTTP and HTTPS.

Instead, this commit extends the existing reconfigure_service_endpoints
method to be invoked on controller not only in the upgrade progress,
calls back this method to update the service endpoints in the above
cases when the puppet classes are applied.

This commit also leverages the greenthreads to execute the
report_config_status, _update_host_config_applied and
_update_runtime_config_status, as those operations don't have
dependencies, improves the runtime config records to be updated in time,
prevents further re-apply of the runtime config in the enrollment
process due to missed DB update caused by service reload.

After this change, those operations take effect much faster than the
puppet approach, 4 minutes can be reduced on each operation, and the
sysinv conductor can be less stressful to deal with deferred manifests
caused by the blocking slow puppet manifest for the service endpoint
update.

Test plan:
1. Passed - update OAM address on simplex.
2. Passed - update OAM address on duplex.
3. Passed - update system between HTTP and HTTPS on simplex.
4. Passed - rehoming a simplex subcloud
5. Passed - enroll a simplex subcloud
6. Passed - parallel 100 host routes creation/deletion successfully
without config out of date alarm.

Closes-bug: 2102767

Signed-off-by: Yuxing Jiang <Yuxing.Jiang@windriver.com>
Change-Id: I3a2847acd684596a48395c69d3ad16b4e59f43de