Reduce service endpoint update time
The time it takes to update keystone endpoints in
"openstack::keystone::endpoint::reconfig" is very long.
This update is to utilize the existing utility to reconfigure
endpoints. This method reclaimed about 7 minutes in the host unlock
command if the management network is changed.

Test Plan:
PASS: Management network reconfig in standalone AIO-SX
PASS: Management network reconfig in DC with AIO-SX subcloud

Closes-Bug: 2103805

Change-Id: I9f1f60a63ab0e4083ca49d21afd5ce46527c5f40
Signed-off-by: Teresa Ho <teresa.ho@windriver.com>