Update Helm repository file using atomic operation
Currently, the process of updating Helm repository files is not atomic,
which can lead to missing or empty files when sysinv is killed or the
machine is forcibly powered off. This commit changes that logic,
allowing the updated file to be written first to a temporary location.
Once the temporary file is written to disk, the original one is
replaced via os.rename, which is an atomic operation that complies
with POSIX requirements.

Test Plan:
PASS: AIO-SX fresh install.
PASS: Upgrade from previous release.
PASS: upload/apply/remove/delete/update metrics-server app

Closes-bug: 2104028

Change-Id: Idce2a8f5400908b0ed0d57b7d1825153256dfe08
Signed-off-by: Igor Soares <Igor.PiresSoares@windriver.com>