Fix deployment model changes in rook-ceph
Some issues were found with the configured rook-ceph backend:

1. Change the deployment model from controller to open and
from open to dedicated, with no OSD checks.

2. Replication set to 3 on AIO-DX with controller deployment model.

3. Replication and min_replication values ​​not having
semantics checked.

4. Default min_replication in open deployment model being set
to integer in database.

Fixes:

1. The 'invalid_model_updates' part has been removed, as it is
possible to switch between the controller and dedicated deployment
model as long as there are no OSDs deployed. This means that the
check when switching from controller to dedicated and from
dedicated to controller, is by checking for the existence of OSDs
in the current deployment model.

2. When it is a DX and the deployment model is 'controller',
the max replication will be 2.

3. Added the checks:
 - capability does not exist in the backend
 - capability with empty value
 - replication and min_replication are integers

4. When min_replication was not defined and the default was used in
the open deployment model, the value was saved as an integer,
causing a problem in DM when reading this column, since it is
expected to be a string. Therefore, the str() casting was added to
solve this issue.

Test Plan:
 - PASS: Create the rook-ceph backend with the
	 controller/dedicated/open deployment model
 - PASS: Modify replication factor with valid and
         invalid values
 - PASS: Modify minimum replication factor with valid
         and invalid values
 - PASS: Modify deployment model
         (switch between deployment models)

* Tests were performed on SX/DX/DX+/STD.

Closes-Bug: 2104919

Change-Id: I32d14fbdf26d7743da19f744c6fed9cd2c2f5185
Signed-off-by: Erickson Silva de Oliveira <Erickson.SilvadeOliveira@windriver.com>