Increase sssd timeout
This commit increases the "timeout" value in the sssd configuration
from the default value of 10s , to 20s.
Timeout setting in "/etc/sssd/sssd.conf" defines the time between
heartbeats for the sssd service. It is used to ensure that the sssd
process is alive and capable of answering requests.
When the "timeout" value is exceeded the sssd watchdog terminates
its child processes and exits.
Increasing the "timeout" mitigates situations when the ssh connection
to a backend system is sluggish, there are networking issues or
other issues related to security policy blocking connection.

Test Plan:
PASS: Verify the sssd domain configuration in "/etc/sssd/sssd.conf"
has been updated with the timeout value of "20".
PASS: Verify that the sssd process is running with all its child
processes.
PASS: Verify that sssd can connect to the local ldap server by creating
a new ldap user and verify the user has been discovered by the platform.
PASS: Verify that ldap commands, like ldapsearch, ldapfinger work
as expected and use ssh to login with the new user.

Closes-Bug: 2105753

Signed-off-by: Carmen Rata <carmen.rata@windriver.com>
Change-Id: I80bbff04a181ed734865f49041128ea5cd7b20b8