Implement inter-app dependency - manual apply action
This change adds to the app_framework the ability to handle
"action" key with value of "apply" from the new dependent_apps
section of the app metadata within the manual processes of
uploading, applying, removing, and deleting an app.

Before applying the target app, the application listed with
the "apply" action in the dependent_apps key in the
application metadata must be fully installed on the platform,
and only then proceed with the installation of the target app.

If the dependent app has dependencies, it must also be installed.

If any app has a circular dependency (meaning that the target app
depends on app x, but app x also depends on the target app),
the apply process is canceled and the target app is left with
applied failed.

This change is a complement to the review:
https://review.opendev.org/c/starlingx/config/+/944261

Here is an example of the new structure supported in the app
metadata:
dependent_apps:
    - name: <app_name.1>
      action: <ignore|warn|error|apply>
    - name: <app_name.2>
      action: <ignore|warn|error|apply>

Test Plan:
PASS: build-pkgs && build-image
PASS: install AIO-SX
PASS: upload, apply, remove and delete metrics-server
PASS: upload, apply, remove and delete NFD
PASS: upload, apply, remove and delete vault

success scenario:
PASS: apply modified vault with the following dependent_app
key:
dependent_apps:
  - name: node-feature-discovery
    version: 25.09-19
    action: apply
  - name: metrics-server
    version: 25.09-58
    action: apply
NFD and metrics-server are successfully installed and then
the vault proceeds to its installation which is completed
successfully.

PASS: apply modified vault with the same dependent_apps above,
but NFD also has dependent_apps. The app dependent NFD is able
to install its own dependencies before moving forward.

failure scenario:
PASS: apply modified vault with the following dependent_app
key:
dependent_apps:
  - name: vault
    version: 25.09-114
    action: apply
  - name: metrics-server
    version: 25.09-58
    action: apply
Vault does not proceed with the installation of dependencies
because it identifies a circular dependency. Vault goes to the
applied failed state and informs in the progress column that a
problem occurred while trying to install the app's
dependencies. It asks the user to check the sysinv.log

PASS: When one of the dependent apps fails the upload or apply
process, this does not stop the operation of the other
dependencies. However, the target app changes to the applied
failed state, since all dependencies must be installed.

Story: 2011351
Task: 51948

Change-Id: Idde0e6f5445e5a2116aa3a03b922dc26d6528b72
Signed-off-by: David Bastos <david.barbosabastos@windriver.com>