Fix recovery process creating wrong .pth files.
A new lifecycle hook for the recovery operation was introduced in the
system by this commit: https://review.opendev.org/c/starlingx/config/+/883974

The lifecycle function requires an app to register the hook. During
the recovery step, the app passed to these lifecycle function was the
version of the app that the system failed to install.
The implication of this is that the activate_app_plugins function was
called for an app version that should no longer exist, leaving the
system at the end of the recovery process with two .pth files.

To solve this problem, this commit removes the incorrect app version
passed to self.app_lifecycle_actions and changes it to the older app
version.

Test plan:
PASS: build-pkgs && build-image.
PASS: AIO-SX fresh install.
PASS: install dell-storage app.
PASS: force an update of dell-storage app to fail and check if the
      system successfully recovers the app to the older version.
PASS: check if the path /var/stx_app/plugins/ has only one version of
      dell-storage.pth file.
PASS: upload and apply metrics-server to check if other application
      plugins were correctly loaded.

Closes-bug: 2106271

Change-Id: Ic3a2f80d227aa9c5b9c7c60d4a136ae4a7938d7f
Signed-off-by: edias <edson.dias@windriver.com>