Fix OSD patch endpoint to allow execution without state transition
This change allows the ceph-sync-osds job, created by the rook-ceph app,
to reset/sync OSDs that are in the 'deleting-with-app' state without
transitioning them to 'configuring-with-app'.

The solution involves updating the logic to use an elif condition
instead of an else, ensuring the OSD state is not modified when it is
set to 'deleting-with-app'.

Test-Plan:
 - PASS: Apply rook-ceph app
 - PASS: Remove OSD and reapply rook-ceph app

Partial-Bug: 2107475

Change-Id: I2baa8cc742dc64ef7911f8cd99111508f832e38f
Signed-off-by: Gustavo Ornaghi Antunes <gustavo.ornaghiantunes@windriver.com>