Check pxeboot IF when assigning pxeboot network
A network configuration issue would appear causing the system
to go into a boot loop if the pxeboot interface is changed
incorrectly. This would prevent the mgmt to get IP from
the primary controller when trying to PXE boot.

This check is to warn the user to ensure that the new
chosen interface is able to execute PXE boot before unlocking.

An error response is not recommended, since the operator may have
already configured the node to PXE boot from the correct interface.

Testing:
- Build successful
- Deploy successful
- Test on 3 different scenarios:
  - On a fresh install where mgmt_mac is not set (no warning)
  - Using an interface with the same MAC as the host MAC (no warning)
  - Using an interface with a different MAC than the host MAC (warning)

Closes-Bug: 2107805

Change-Id: I6a900933e6f47f894e267fbb200280b6ce415ee4
Signed-off-by: Mohammad Issa <mohammad.issa@windriver.com>