Fix deprecated mgmt IP when sharing IPv6 vlan
When mgmt and cluster-host networks are on the same IPv6 vlan, the mgmt
IP was deprecated causing a worker node from being unlocked.
In this case, the cluster-host IP should have been deprecated and mgmt
IP should be undeprecated.
This commit is to ensure that mgmt IP is undeprecated when mgmt network
is sharing a vlan with another network.

Test Plan:
PASS: install IPv6 AIO-DX with compute
PASS: install IPv4 standard with mgmt and cluster-host on same
      interface

Closes-Bug: 2107505

Change-Id: I3d695d63114849a46a5189f447849aaa0b870868
Signed-off-by: Teresa Ho <teresa.ho@windriver.com>