Limit monitor creation to 5 with system host-fs
The platform allocates Ceph monitors on all hosts that have a "ceph"
host-fs with the "monitor" capability. However, this is also used to
define the number of Ceph mgr to be created, and that is currently
limited to 5 by Rook Ceph. If there are more than that, the rook-ceph
application apply will return the error "spec.mgr.count in body
should be less than or equal to 5". This commit limits the number of
host-fs with "monitor" capability to 5, to avoid creating more than
5 Ceph mgr. This also limits the number of monitors to 5.

Test Plan:
 - PASS: In a system with no monitors configured, add 5 monitors
         successfully and get an error when trying to add the sixth
         one with "system host-fs-add"
 - PASS: In a system with 5 monitors, try to add the Monitor role to
         a sixth host and get an error with "system host-fs-modify"
 - PASS: In a system with 5 monitors, remove the Monitor role from
         any of the hosts and add it to a sixth host with "system
         host-fs-modify"
 - PASS: In a system with no monitors, add 3 of them, wait for the
         application apply to finish and try to add 3 more, getting
         an error when adding the sixth monitor
 - PASS: Fresh install with a designer ISO
 - PASS: Swact controllers
 - PASS: Loc/Unlock hosts
 - PASS: Reboot hosts one at a time

Closes-bug: 2109753

Change-Id: I449d5f5f7dcbcff3a080e50e44a1906611e41bea
Signed-off-by: Raphael Santos <raphael.dasilvasantos@windriver.com>