Missing dnsmasq addn_conf file
In an unexpected rare case where the system fails to resolve
"controller.internal" due to the missing dnsmasq.addn_conf file,
preventing dnsmasq from starting.

This error was encountered after unlocking the controller and
did not allow system commands to execute.

In a case where puppet was unable to properly mount the conf file,
the call to _generate_dnsmasq_conf_file() would ensure that the conf
file exists before unlocking. This function is periodically executed,
guaranteeing the file is available once and ignored if present.

Add a flag to ignore the dnsmasq reboot in
generate_dnsmasq_conf_file() since it will be done
in generate_dnsmasq_hosts_file()

Testing:
- Build successful
- Deploy successful
- Test on affected system:
  - Ensure the dnsmasq.add_conf file exists after bootstrap
  - Ensure controller unlocks successfully
  - Ensure dnsmasq was able to start successfully
  - Add a dns service parameter using CLI to ensure functionality

Closes-Bug: 2109381

Change-Id: I04d5806356efb38ea0e54d394d0a04182611b417
Signed-off-by: Mohammad Issa <mohammad.issa@windriver.com>