Prevent unlock when apps are uploading/removing
Prevent host unlock when applications are in uploading (simplex-only) or
removing states. This will avoid applications from transitioning to
upload-failed or remove-failed states when the user inadvertently issues
a system host-unlock command when applications are uploading or
removing.

Test plan:
AIO-SX:
PASS: fresh install
PASS: Attempt to unlock host when oidc-auth-apps is in uploading state.
      Confirm that unlock is blocked.
PASS: Attempt to unlock host when oidc-auth-apps is in removing state.
      Confirm that unlock is blocked.
PASS: Attempt to unlock host when oidc-auth-apps is in uploaded state.
      Confirm that unlock is allowed.

Multi-node:
PASS: fresh install
PASS: Attempt to unlock host when oidc-auth-apps is in uploading state.
      Confirm that unlock is allowed.
PASS: Attempt to unlock host when oidc-auth-apps is in removing state.
      Confirm that unlock is blocked.
PASS: Attempt to unlock host when oidc-auth-apps is in uploaded state.
      Confirm that unlock is allowed.

Closes-bug: 2109534

Change-Id: I6b41846e0c94d5d79bedefa9492076149e70158f
Signed-off-by: Igor Soares <Igor.PiresSoares@windriver.com>