Add include parameter to StrongSwan config
This commit implements the addition of a include parameter to the
strongswan config file in order to include multiple conf file per
the specification of their filepath.

Test Paln:
PASS: Build ipsec-auth and sysinv packages.
PASS: Build, install and deploy a DX system with a worker and a storage
      nodes containing these changes. Observe that swanctl.conf file is
      generated with the all included files at the EOF.
PASS: Install ipsec-policy-operator app, create an IPsecPolicy and wait
      until IPsec Policy Agent generates k8s-node.conf files for each
      kubernetes node. Reload IPsec connections and observe that conns
      under k8s-nodes.conf file are reloaded.

Story: 2011127
Task: 52163

Depends-On: https://review.opendev.org/c/starlingx/stx-puppet/+/948583

Change-Id: I1b682a8e7bb80bdc2f231db0048754286aac48ad
Signed-off-by: Manoel Benedito Neto <Manoel.BeneditoNeto@windriver.com>