Long-term improvement for the application recovery lifecycle.
The bug caused by the activation of the wrong version of a plugin
during an application recovery was fixed using a conservative
approach by this commit:
https://review.opendev.org/c/starlingx/config/+/946383

That entailed loading the previous version of the plugins for the
recovery lifecycles, given that the old version was more likely to be
working compared to the failing updated version.
However, this approach did not allow app developers to use the updated
recovery code from the newer version of the application.

This commit aims to provide a long-term solution that safely executes
the recovery code from the updated version of the application.

Test plan:
PASS: build-pkgs && build-image.
PASS: AIO-SX fresh install.
PASS: install metrics-server app.
PASS: force an update of metrics-server app to fail and check if the
      system successfully recovers the app to the older version.
PASS: check if the path /var/stx_app/plugins/ has only one version of
      metrics-server.pth file.
PASS: upload and apply portieris app to check if other application
      plugins were correctly loaded.

PASS: install metrics-server app.
PASS: force an metrics-server update and a lifecycle_recover to fail,
      and check if the recover process finishes successfully.

Closes-bug: 2109840

Change-Id: Ie87e874994e67c1557f7f0583553a53b9a468967
Signed-off-by: edias <edson.dias@windriver.com>