Pass config_uuid to report function and increase retry time
When multiple runtime manifests are requested to be
executed in a short period of time, a potential issue
might happen where a runtime manifest reports success
but using the config_uuid from another runtime manifest.
This is a consequence of multiple hieradata updates and the
async nature of the runtime manifests execution.

There is another issue found in slower or more loaded systems,
where a small number of manifests may take more than 10 minutes
to execute, leading the system to retry the manifest since
it has not yet reported success in a 10 minute period, and
this can create a chain of unnecessary retries of runtime
manifests, leading to a much longer system stabilization time.

This commit passes the config_uuid as an optional parameter to
the report function, instead of relying in the config_uuid value
from the hieradata; when the optional parameter is not passed,
the value from the hieradata is used. Also, this commit increases
the wait-to-retry period of runtime manifests to 20 minutes, to
avoid longer stabilization times in slower systems.

There will be additional work to solve additional runtime
manifests concurrence issues in follow-up commits.

Test Plan
PASS: trigger multiple runtime manifests at the same moment,
      verify the correct config_uuid is reported by each
      runtime manifest
PASS: AIO-SX subcloud e2e upgrade, verify the correct config_uuid
      is reported by the multiple runtime manifests triggered
      near the end of the upgrade process
PASS: run stress test with multiple runtime configs in parallel,
      verify no alarm is stuck after it
PASS: install/bootstrap/unlock AIO-SX

Closes-bug: 2109860

Change-Id: I01fa182119d3b5e1dd5787a1f7ff2c0aa37cf140
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>