Compute order of apps for automatic process
This change prepares the environment for a new strategy for ordering
application installations. The recompute_app_evaluation_order
function now returns a dictionary with 3 keys: dependent_apps, class
and independent_apps.

1 - the "dependent_apps" key contains a list of apps that use
"dependent_apps" in their metadata. This key in the apps of the
metadata is used to list other apps that need to be installed first

2 - The "class" key contains a dictionary of lists organized as
follows:
{
  critical: [],
  storage: [],
  discovery: [],
  optional: [],
  reporting: []
}
Each app that uses the "class" key will have its respective list added
to this dictionary.

3 - And finally, the "independent_apps" key are apps that do not use
either "dependent_apps" or "class" in their metadata.

The scope of this change is only to compute this new list using these
new keys (dependent_apps, class) supported in the metadata.yaml of the
applications. A new task 52102 from storyboard 2011351 will be
responsible for using this new format. The automatic installation of
apps present in this dictionary will be performed in series or in
parallel, depending on where it is located.

When the entire feature from storyboard 2011351 is complete, the "after"
key from the application metadata currently used to compute the order
of the apps will be removed from all apps loaded by the platform.
However, the previous logic will be maintained and if the "after" key is
identified in the metadata of any application, the previous logic will
be called to compute the ordered list of apps.

Test Plan:
PASS: build-pkgs && build-image
PASS: fresh install AIO-SX
PASS: upgrade starlingx 11 to master
PASS: If you remove any app using the after key
      the recompute_app_evaluation_order_by_after_key function will be
      called and a list of apps will be returned without breaking
      backwards compatibility.
PASS: If no app is using the key after, the
      recompute_app_evaluation_order function will be called and the
      dictionary with the list of apps will be returned.
PASS: If any app that uses the "class" key also has the "dependent_apps"
      key, only the "dependent_apps" key will be considered.

Story: 2011351
Task: 52100

Change-Id: I311141e8aad800a36208d5764c2d5985bca6c698
Signed-off-by: David Bastos <david.barbosabastos@windriver.com>