Use internal endpoint in keystone client
Keystone client instances in sysinv-conductor are defaulting to the
keystone admin endpoint. This causes issues during subcloud bootstrap
in some instances, as haproxy changes to listen to 5001 port for the
keystone admin endpoint after the first unlock, and sysinv is not
able to reconfigure the endpoint because it's still using the 5000
port for it.

The internal endpoint always remain using the 5000 port, so it can
be used for the instances of the client in every scenario.

Test plan:
PASS: Bootstrap AIO-SX
PASS: Bootstrap DC Systemcontroller + SX subcloud
      Confirmed that subcloud is online and in sync
PASS: Bootstrap SX subcloud w/ admin network
      Confirmed that subcloud is online and in sync

Closes-bug: 2110481

Change-Id: I2ae86bc1da3a0a55edfe2f62b19581a7ef52cf3d
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>