Ensure admin network has an associated interface
The system should not allow the user to unlock the system if one
configures the admin network but doesn't associate it with
an interface.

The case was that the admin network was conmfigured in the
bootstrap-values.yml but was never assigned to an interface
prior to first unlock.

This commit checks if an admin network is configured and ensures
that it is associated with an interface.

Testing:
- Build successful
- Deploy successful
- Deployed a subcloud with admin network configured
  but without an associated interface (block unlock)
- Deployed a subcloud with admin network configured
  and associated with an interface (allow unlock)

Closes-Bug: 2111116

Change-Id: Ie73a94a66af3140831ac572592c3e7554817a737
Signed-off-by: Mohammad Issa <mohammad.issa@windriver.com>