Apply dependencies during application automatic operations
This change is a continuation of the review:
https://review.opendev.org/c/starlingx/config/+/948802/13

Basically, it consists of using the new data structure stored in
the ordered_apps key within the apps_metadata attribute of the
AppOperator class. Previously, the ordered_apps key was a list, but
now it becomes a dictionary, in case the "after" key is no longer
present in the apps loaded by the platform. The apps_metadata
attribute stores a cache of the metadata of all applications stored
in the /usr/local/share/applications/helm/ folder.

This way, the determine_apps_reapply_order function now takes into
account the dictionary and no longer the list present in
ordered_apps. The previous behavior is being maintained for backward
compatibility, but in a new release it should be removed.

Another improvement added to this change is the option to wait for
an app dependency to finish uploading in progress or applying in
progress. This scenario can happen if 2 or more apps that are being
installed in parallel have the same app as a dependency.

With the possibility of apps that have other apps as dependencies
being able to be installed/updated in parallel, this change also
removes the lock on the auto_apply and auto_update functions of
updating only one application at a time.

Test Plan:
PASS: build-pkgs && build-image
PASS: fresh install AIO-SX
PASS: upgrade starlingx 11 to master
PASS: Add applications with common dependencies and perform the
      apply of these apps in parallel manually.\
PASS: Add metrics-server and vault with the desired state of apply
      both depending on the kubervirt application and observe, as
      soon as the platform is installed, these apps being installed
      in parallel automatically.
PASS: Upload, remove, apply e delete metrics-server app
PASS: Upload, remove, apply e delete vault app
PASS: Upload, remove, apply e delete NFD app

Story: 2011351
Task: 52216

Change-Id: I36fa75351ba61cd11a17e41280ece14823cc2552
Signed-off-by: David Bastos <david.barbosabastos@windriver.com>